#59: Accuracy
A base Accuracy class to be added to every attack, that is inherited by its subclasses.  These subclasses say which defense is targeted by the attack.
