+++ Assets/Scripts/View Model/Actor/Abilities/HitRate/AccuracyDeflection.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d8668d187aefb444fab1061e864929eb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
