+++ Assets/Scripts/View Model/Actor/Abilities/HitRate/AccuracyFull.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class AccuracyFull : Accuracy
+{
+    public override int Calculate(Tile tile)
+    {
+        if (AutomaticMiss(tile))
+        {
+            return 0;
+        }
+
+        return int.MaxValue;
+    }
+}
