+++ Assets/Scripts/View Model/Actor/Abilities/HitRate/AccuracyFull.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ea3d0dae13a464e47a47b0e4300f0b83
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
