+++ Assets/Scripts/View Model/Actor/Abilities/HitRate/AccuracyResolve.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d61b59dab8d0fbe44bd7ea7b7a459486
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
